1.扩展协议，编号修改为2个字节，第一个字节表示某种业务，第二个字节表示协议编号 2.修改了套接字发送的接口，现在只保留通过channel和tag发送，并且忽略了tcp和ucp差异

+++ Net/ISocketHandler.cs
@@ -51,7 +51,7 @@ namespace Kakegurui.Net
         /// <summary>
         /// 发送协议编号
         /// </summary>
-        public int ProtocolId { get; set; }
+        public ushort ProtocolId { get; set; }
 
         /// <summary>
         /// 发送时间戳
